The stream atom feed is currently timing out in production. There is a bunch of different stuff going on here (including the fact that the atom feed makes requests to our API by going out to the public internet and back in rather than directly to the service running on the same host...) but benchmarking suggests that the two biggest sources of slowness are:

1. JSON serialisation (specifically "pretty" output)
2. Making two identical queries to ElasticSearch instead of one...

In addition, I think 1000 annotations in the atom feed is probably a bit much -- an atom feed approaching a megabyte seems excessive. I've dropped the default number of annotations to 100.

With these three changes we go from about 1 request per second...

```
$ wrk -c1 -d10s -t1 'http://localhost:5000/stream.atom'
Running 10s test @ http://localhost:5000/stream.atom
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   727.94ms   31.91ms 772.22ms   53.85%
    Req/Sec     1.00      0.00     1.00    100.00%
  13 requests in 10.09s, 10.79MB read
Requests/sec:      1.29
Transfer/sec:      1.07MB
```

...to about 13:

```
$ wrk -c1 -d10s -t1 'http://localhost:5000/stream.atom'
Running 10s test @ http://localhost:5000/stream.atom
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    71.99ms   17.25ms 159.24ms   95.10%
    Req/Sec    13.58      5.06    20.00     57.29%
  139 requests in 10.03s, 12.61MB read
Requests/sec:     13.86
Transfer/sec:      1.26MB
```