@tilgovi 2015-07-03
Other than my comment about the 'json' renderer this is great. Merge it after you fix that.